Create dedicated test_utils
