+++ sockpuppet/test_utils/__init__.py
@@ -1,0 +1,1 @@
+from .factory import reflex_factory # noqa
