--- sockpuppet/test_utils/test_app/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class TestAppConfig(AppConfig):
-    name = 'test_app'
